% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aggregate_flows.R
\name{aggregate_flows}
\alias{aggregate_flows}
\title{Aggregate flows to specified geography}
\usage{
aggregate_flows(geography = c("state", "county"))
}
\arguments{
\item{geography}{one of \code{c("state", "county")} specifying the level of
aggregation desired.}
}
\value{
a \code{tbl_df(data_frame)} with the \code{ctpp_flows} flow data aggregated
 to the specified geographic level.
}
\description{
Aggregate flows to specified geography
}
\details{
This function uses \code{dplyr}'s efficient interfaces to aggregate
the flows to a specified geographic level. The flow estimates are summed,
and the margins of error are aggregated using Census' recommended formula,
\deqn{\sqrt{\sum_i (x_i^2)}}
}
\examples{
aggregate_flows("state")
}

