% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fips_functions.R
\name{get_county}
\alias{get_county}
\title{Get the county-level FIPS code}
\usage{
get_county(fips, with_state = FALSE)
}
\arguments{
\item{fips}{A character vector of complete FIPS strings.}

\item{with_state}{Include the 2-digit state code? Default
\code{with_state = FALSE}}
}
\value{
a character vector with the 3-digit county code, unless
  \code{with_state = TRUE} in which case a 5-digit state and county code
  concatenation.
}
\description{
\code{get_county} returns the 3-digit county FIPS code from a complete FIPS
string.
}
\examples{
get_county(ctpp_flows$residence[1:10])
get_county(ctpp_flows$residence[1:10], with_state = TRUE)
}
\seealso{
\code{\link{get_state}}, \code{\link{get_tract}}
}

