% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fips_functions.R
\name{get_state}
\alias{get_state}
\title{Get the state-level FIPS code}
\usage{
get_state(fips)
}
\arguments{
\item{fips}{A character vector of complete FIPS strings.}
}
\value{
a character vector with the 2-digit state code.
}
\description{
\code{get_state} returns the 2-digit state FIPS code from a complete FIPS
string.
}
\examples{
get_state(ctpp_flows$residence)
}
\seealso{
\code{\link{get_county}}, \code{\link{get_tract}}
}

