% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fips_functions.R
\name{get_tract}
\alias{get_tract}
\title{Get the tract-level FIPS code}
\usage{
get_tract(fips, with_state = FALSE)
}
\arguments{
\item{fips}{A character vector of complete FIPS strings.}

\item{with_state}{Include the 5-digit state and county code? Default
\code{with_state = FALSE}. This returns the existing FIPS code.}
}
\value{
a character vector with the 6-digit county code, unless
  \code{with_state = TRUE} in which case a 11-digit state, county, and tract
  code concatenation.
}
\description{
\code{get_tract} returns the 6-digit tract FIPS code from a complete FIPS
string.
}
\examples{
get_tract(ctpp_flows$residence[1:10])
get_tract(ctpp_flows$residence[1:10], with_state = TRUE)
}
\seealso{
\code{\link{get_county}}, \code{\link{get_state}}
}

